cargo: fix cargo-subcommand argument passing
Cargo treats binaries on your path named "cargo-X" as subcommands and
exec's "cargo-X" when calling "cargo X". Unfortunately when this feature
was implemented the incorrect arguments were passed to the subcommand
resulting in "cargo-X" being invoked with the first argument being the
name of the subcommand (i.e. "cargo-X X <remaining args>) instead of
like how git handles subcommands (i.e "cargo-X <remaining args>").

In order to account for this (and still allow for the binary being
invoked directly) preprocess the args, removing the first arg if it
matches the name of the cargo subcommand, before passing them to
StructOpt.