[graph] add a ReversedDirected adapter
This is similar to `petgraph::visit::Reversed`, except that it has implementations for `IntoEdges` and `IntoEdgesDirected`. This adapter will be used in upcoming patches.

Due to an inconsistency in `petgraph` (https://github.com/petgraph/petgraph/issues/292), this adapter can only be used for directed graphs. Hopefully a future version of `petgraph` will fix this inconsistency.