[visit] implement a custom DotFmt operator
This is adapted from the upstream `Dot` support in petgraph, but it uses a visitor.
This allows for a lot of customization in what node and edge labels get printed.

This is written in a pretty generic style, and it works over all structures
that implement the requisite. This could potentially be upstreamed to petgraph
if found useful.