[guppy] upgrade to petgraph 0.5
The one code change required was updating the definition of `reachable_map` to
reflect that the dfs algorithm is different from before. With the new DFS
algorithm, if a node is already in the visit map it will no longer be
traversed -- this means that visit_map should no longer be pre-seeded with the
roots.

Also, disable petgraph features we don't use for quicker compile times.

This is not a breaking change because no petgraph interfaces are exposed to
consumers of guppy.