[graph] re-do platform-specific statuses
I've been a bit unsatisfied by the current design of the way `EnabledStatus` etc work.

* The current design always requires evaluation against a particular platform. We've made this easier by providing the current platform as default, but that is error-prone as #111 points out.
* The current design also doesn't have a way to distinguish between dependencies that are enabled regardless of platform and those that are enabled on a specific platform (but might be disabled on others).
* The API also isn't very orthogonal -- in particular, feature graph resolution caused me to have to introduce the `PlatformStatus` API anyway, and we should try and make that API also work for the package graph.

This new design is an answer to all of those concerns. It introduces three data structures:
* `EnabledStatus`, which is a pair of (required, optional) `PlatformStatus` instances.
* `PlatformStatus`, which is always, never or platform-dependent.
* `PlatformEval`, which represents a set of target specs to evaluate against a platform.

The API is also designed to provide information at the level of granularity one desires. For example, if one isn't interested in distinguishing between enabled-everywhere and enabled-on-platform, one can just call `PlatformStatus::enabled_on`. If that information is required, users can always match on the `PlatformStatus`.

This also cleans up many issues that were somewhat annoying to me -- for example, `cargo-guppy`'s `--include-target` flag only checked for the normal edge. With this commit, `cargo-guppy` combines  the `--include-target`, `--include-dev` and `--include-build` flags in a way that feels very natural.

Closes #110.
Closes #111.