[resolve] allow PackageResolver to be &mut self instead of &self
`EdgeFiltered` only works for `Fn` instances, so this requires a bit of
manual filtering. It's not too bad, though.

This is more correct in the sense that it is the right level of generality,
and also I'd also like to exploit this in some upcoming work and want to
avoid RefCell.