[guppy] return packages in cycles in non-dev order
Before this commit, packages in cycles could be returned in arbitrary
order. But there is a natural order to return them in: non-dev order.

This is not a breaking change because it's just an additional API constraint
on guppy that didn't exist before.

Constructing cycles does become slower as a result, but it's still
pretty fast. There's likely more scope for optimization in the future.

make_cycles             time:   [118.21 us 119.18 us 120.11 us]
                        change: [+77.496% +78.481% +79.512%] (p = 0.00 < 0.05)