cargo: fix cargo subcommand argument passing on Windows
External tools which extend Cargo with custom subcommands get called with the name of the subcommand as the first argument.

If it is desired for the external tool to also be callable separately, it is useful to remove this first argument if it exists, so no repetition of the subcommand name takes place.

The code which makes sure this happens works like a charm on Linux, but on Windows, executables usually carry the ".exe" file name extension.

This commit fixes the above issue, so now cargo-guppy can also be called as a custom Cargo subcommand on Windows.