[hakari] accept platform triples rather than more complex information
In the future we may want to allow platforms to specify `cfg()`
expressions, then resolve them with an SMT solver.

Supporting full-fidelity platforms seems like the wrong choice given the
fact that we need to output `[target.$TRIPLE.dependency]` lines. Also,
target triples are supported as both platforms and in place of `cfg()`,
so they're the common subset across both.