[guppy] handle inactive, optional dependencies better
There was a slight issue with the way we were handling inactive,
optional dependencies. This PR should fix that.

Also expand the inside-outside tests to cover both the v1 and v2
resolvers; requires copying the fixture into a temp dir so that the
'resolver = "1"' or "2" can be added to it.

hakari output changes slightly -- this is a bugfix so it's a
non-breaking change.

See the comments in this PR for more.